New uploader, with no cache
The old uploader had a cache to remember which files were stored to S3.

And on 2016-02-27, the production server didn't upload anything to S3
until we wiped out the entire `cache/` directory (which includes the
uploader's cache of which files are stored in S3).

We checked that the cached election_day-2016-02-27 was correct. And
`production-server` removes all files from `dist/` each time it runs.
That suggests the issue might have been with the cache.

No cache, no problem? We shall see.